% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{cran.stats}
\alias{cran.stats}
\title{Get summary statistics for \code{R} packages available on Github and CRAN}
\usage{
cran.stats(spec_author = c("hornik", "leisch"))
}
\arguments{
\item{spec_author}{an optional character vector with the desired author list for being displayed}
}
\description{
Get summary statistics for \code{R} packages available on Github and CRAN
}
\examples{
\dontrun{
cran.stats(spec_author = c("hornik", "leisch"))
}
}

