% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgs_repos_members_files_LB.R
\name{create.file.content.on.github}
\alias{create.file.content.on.github}
\title{API Wrapper for file creation}
\usage{
create.file.content.on.github(user_name, repo_name, file_path, file_name,
  file_string_content, api_create_message)
}
\arguments{
\item{user_name}{the user name}

\item{repo_name}{the name of the repo}

\item{file_path}{the file path}

\item{file_name}{the file name}

\item{file_string_content}{the file content}

\item{api_create_message}{the message with the status on file creation}
}
\description{
API Wrapper for file creation
}

