% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgs_repos_members_files_LB.R
\name{create.file}
\alias{create.file}
\title{Create a new file}
\usage{
create.file(owner, repo, path, content, ctx = get.github.context())
}
\arguments{
\item{owner}{the repo owner (user, org, etc)}

\item{repo}{the name of the repo}

\item{path}{the file path}

\item{content}{the content(?)}

\item{ctx}{the github context object}
}
\description{
Create a new file
}

