% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{dcf.search.parser}
\alias{dcf.search.parser}
\title{Parse DCF encoded metadata retrieved via \code{search.code}}
\usage{
dcf.search.parser(search.code.results, print_item = FALSE,
  error_to_results = TRUE)
}
\arguments{
\item{search.code.results}{results from \code{search.code}}

\item{print_item}{if true, then print retrieved items}

\item{error_to_results}{if true, then print information on errors}
}
\value{
a list with processed data
}
\description{
Parse DCF encoded metadata retrieved via \code{search.code}
}
\examples{
\dontrun{

# search for R packages
r_pack_search = 'Package Title Version Description Author
				filename:"DESCRIPTION" path:"/"'
spec_search_term = "yaml"
sr = search.code(paste(spec_search_term, r_pack_search), per_page = 10)
dcf_top = dcf.search.parser(sr)
}
}

