% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgs_repos_members_files_LB.R
\name{get.all.orgs}
\alias{get.all.orgs}
\title{Get all organizations}
\usage{
get.all.orgs(..., ctx = get.github.context())
}
\arguments{
\item{...}{extra parameters, see \url{http://developer.github.com/v3/repos/}}

\item{ctx}{the github context object}
}
\value{
information on all organizations
}
\description{
Get all organizations
}

