% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgs_repos_members_files_LB.R
\name{get.api.ratelimit}
\alias{get.api.ratelimit}
\title{Get information on the remaining API ratelimit}
\usage{
get.api.ratelimit(org_name)
}
\arguments{
\item{org_name}{the organization name}
}
\description{
Get information on the remaining API ratelimit
}

