% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgs_repos_members_files_LB.R
\name{get.organization.public.members.fulllist}
\alias{get.organization.public.members.fulllist}
\title{Get information on public members of given organization}
\usage{
get.organization.public.members.fulllist(org_name)
}
\arguments{
\item{org_name}{the organization name}
}
\value{
list with information on public members
}
\description{
Get information on public members of given organization
}

