% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgs_repos_members_files_LB.R
\name{get.organization.public.members.pagination}
\alias{get.organization.public.members.pagination}
\title{Get list of public members of given organization}
\usage{
get.organization.public.members.pagination(..., org,
  ctx = get.github.context())
}
\arguments{
\item{...}{extra parameters, see \url{http://developer.github.com/v3/repos/}}

\item{org}{the given organization}

\item{ctx}{the github context object}
}
\value{
list of public members
}
\description{
Get list of public members of given organization
}

