% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgs_repos_members_files_LB.R
\name{get.organization.public.repos.fulllist}
\alias{get.organization.public.repos.fulllist}
\title{Get list of public repositories of given organization}
\usage{
get.organization.public.repos.fulllist(org_name)
}
\arguments{
\item{org_name}{the organization name}
}
\value{
list of public repositories
}
\description{
Get list of public repositories of given organization
}

