% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgs_repos_members_files_LB.R
\name{get.repositories.full.description}
\alias{get.repositories.full.description}
\title{Get information on repositories of given organization}
\usage{
get.repositories.full.description(org_name, public_repos_nr)
}
\arguments{
\item{org_name}{the organization name}

\item{public_repos_nr}{the number of public repositories}
}
\value{
list with information on repositories
}
\description{
Get information on repositories of given organization
}

