% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{qnet.stats}
\alias{qnet.stats}
\title{Get summary statistics for Quantlets available on Github}
\usage{
qnet.stats(spec_editor = c("borke", "bykovskaya"),
  json_path = "http://quantlet.de/data/qlets_github_ia_utf8.json")
}
\arguments{
\item{spec_editor}{an optional character vector with the desired author list for being displayed}

\item{json_path}{an optional character with the url for json file with all validated Quantlets available on QuantNetXploRer}
}
\description{
Get summary statistics for Quantlets available on Github
}
\examples{
\dontrun{
cran.stats(spec_editor = c("borke", "bykovskaya"))
}
}

